Reorganize SEP-9 data and remove unneeded enums.
